requirements.txt
absl-py==1.0.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.12b0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.3
executing==0.8.2
filelock==3.4.2
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
gdown==4.2.0
google-api-core==2.4.0
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
h5py==3.6.0
httplib2==0.20.2
idna==2.8
importlib-metadata==4.10.0
ipykernel==4.6.1
ipython==8.0.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mediapipe==0.8.9.1
mistune==0.8.4
mypy-extensions==0.4.3
myutils-python @ https://github.com/jordiyapz/myutils-python/releases/download/v0.1.0/myutils_python-0.1.0-py3-none-any.whl
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
numpy==1.22.0
oauth2client==4.1.3
oauthlib==3.1.1
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
portpicker==1.2.0
prompt-toolkit==3.0.24
protobuf==3.19.3
pure-eval==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
PyDrive==1.3.1
Pygments==2.11.2
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
PyYAML==6.0
pyzmq==22.3.0
questionary==1.10.0
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.8
simplegeneric==0.8.1
six==1.12.0
soupsieve==2.3.1
stack-data==0.1.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
testpath==0.5.0
toml==0.10.2
tomli==1.2.3
tornado==4.5.3
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
# Editable Git install with no remote (yes-capture==1.0.0)
-e c:\users\jordi\experiments\pose_tangan
zipp==3.7.0
